+++ src/components/HeartButton/index.js
@@ -3,15 +3,12 @@ import { useSelector, useDispatch } from "react-redux"
 
 import { newFavorite, removeFavorite } from "../../store/User/actions"
 import { selectFavorites } from "../../store/User/selectors"
-import { selectUser } from "../../store/User/selectors"
 import "./index.css"
 
 export default function HeartButton(props){
     const dispatch = useDispatch()
     const favorites = useSelector(selectFavorites)
     // console.log("favorites test", favorites)
-    const user = useSelector(selectUser)
-    // console.log("user test", user.token)
     const product = props.data
     // console.log("product prop test", product)
